scripts/requirements.txt
graphviz
isodate
Lifetimes
lightgbm
matplotlib
numpy
scikit-learn
scipy
seaborn
sidetable
simple-colors
sklearn
sklearn-pandas
pandasql
pydotplus
ipykernel
statsmodels
azure-storage-blob
ipython-sql
pyodbc